+++ src/selection_indicator/style/mod.rs
@@ -9,7 +9,7 @@ use crate::{interaction::InputState, selection_indicator::Insets, theme::Theme};
 pub mod animated_triangle;
 pub mod border;
 pub mod line;
-// pub mod rectangle;
+pub mod rectangle;
 pub mod triangle;
 
 // Re-export the styles themselves to make them easier to use.