+++ src/selection_indicator/style/rectangle.rs
@@ -1,10 +1,9 @@
 use embedded_graphics::{
-    prelude::{DrawTarget, Point, Size},
-    primitives::{ContainsPoint, Primitive, PrimitiveStyle, Rectangle as RectangleShape},
-    transform::Transform,
+    pixelcolor::BinaryColor,
+    prelude::DrawTarget,
+    primitives::{Primitive, PrimitiveStyle, Rectangle as RectangleShape},
     Drawable,
 };
-use embedded_layout::prelude::{horizontal::LeftToRight, vertical::Center, Align};
 
 use crate::{
     interaction::InputState,
@@ -12,6 +11,7 @@ use crate::{
         style::{interpolate, IndicatorStyle},
         Insets,
     },
+    theme::Theme,
 };
 
 #[derive(Clone, Copy)]
@@ -19,24 +19,35 @@ pub struct Rectangle<C = BinaryColor> {
     color: C,
 }
 
+impl<C> Rectangle<C> {
+    pub fn new(color: C) -> Self {
+        Self { color }
+    }
+}
+
 impl<C> IndicatorStyle for Rectangle<C>
 where
-    C: Copy,
+    C: Theme,
 {
-    type Shape = Arrow<C>;
+    type Color = C;
+    type Shape = RectangleShape;
     type State = ();
 
-    fn padding(&self, _state: &Self::State, height: i32) -> Insets {
+    fn padding(&self, _state: &Self::State, _height: i32) -> Insets {
         Insets {
-            left: height / 2 + 1,
+            left: 2,
             top: 0,
             right: 0,
             bottom: 0,
         }
     }
 
-    fn shape(&self, _state: &Self::State, bounds: RectangleShape, fill_width: u32) -> Self::Shape {
-        Arrow::new(bounds, fill_width, self.color)
+    fn shape(&self, _state: &Self::State, bounds: RectangleShape, _fill_width: u32) -> Self::Shape {
+        bounds
+    }
+
+    fn color(&self, _state: &Self::State) -> <Self::Color as Theme>::Color {
+        self.color.selection_color()
     }
 
     fn draw<D>(
@@ -58,7 +69,9 @@ where
 
         let shape = self.shape(state, display_area, fill_width);
 
-        shape.draw(display)?;
+        shape
+            .into_styled(PrimitiveStyle::with_fill(self.color.selection_color()))
+            .draw(display)?;
 
         Ok(shape)
     }